test: split oversized e2e helpers
