+++ crates/astrid-capsule/src/engine/wasm/host/elicit.rs
@@ -622,395 +622,10 @@ mod tests {
     }
 }
 
-// ---------------------------------------------------------------------------
-// Wait-loop integration test: drive the real `elicit()` host fn against a live
-// `EventBus` and prove the responder-principal enforcement end to end — a reply
-// from the WRONG principal does NOT unblock the waiter, while a reply from the
-// MATCHING principal does. The pure-helper test above guards the decision; this
-// one guards the loop that consumes it (the cross-principal reply must be
-// skipped and the wait must continue on its remaining budget).
-// ---------------------------------------------------------------------------
 #[cfg(test)]
-mod wait_loop_tests {
-    use std::time::Duration;
+#[path = "elicit/wait_loop_tests.rs"]
+mod wait_loop_tests;
 
-    use crate::engine::wasm::bindings::astrid::elicit::host::{
-        ElicitRequest, ElicitResponse, ElicitType, ErrorCode, Host as ElicitHost,
-    };
-    use crate::engine::wasm::host_state::LifecyclePhase;
-    use crate::engine::wasm::test_fixtures::minimal_host_state;
-    use astrid_events::AstridEvent;
-    use astrid_events::ipc::{IpcMessage, IpcPayload, Topic};
-    use uuid::Uuid;
-
-    fn text_request(key: &str) -> ElicitRequest {
-        ElicitRequest {
-            kind: ElicitType::Text,
-            key: key.to_string(),
-            description: "Enter a value".to_string(),
-            options: None,
-            default_value: None,
-        }
-    }
-
-    /// Publish an `ElicitResponse` for `request_id` stamped with `principal`.
-    fn publish_response(
-        bus: &astrid_events::EventBus,
-        request_id: Uuid,
-        principal: &str,
-        value: &str,
-    ) {
-        let topic = Topic::elicit_response(request_id);
-        let msg = IpcMessage::new(
-            topic,
-            IpcPayload::ElicitResponse {
-                request_id,
-                value: Some(value.to_string()),
-                values: None,
-            },
-            Uuid::nil(),
-        )
-        .with_principal(principal);
-        bus.publish(AstridEvent::Ipc {
-            message: msg,
-            metadata: astrid_events::EventMetadata::default(),
-        });
-    }
-
-    /// Publish an `ElicitResponse` for `request_id` with BOTH `value` and
-    /// `values` `None` — the host's cancellation sentinel — stamped with
-    /// `principal`.
-    fn publish_cancel(bus: &astrid_events::EventBus, request_id: Uuid, principal: &str) {
-        let topic = Topic::elicit_response(request_id);
-        let msg = IpcMessage::new(
-            topic,
-            IpcPayload::ElicitResponse {
-                request_id,
-                value: None,
-                values: None,
-            },
-            Uuid::nil(),
-        )
-        .with_principal(principal);
-        bus.publish(AstridEvent::Ipc {
-            message: msg,
-            metadata: astrid_events::EventMetadata::default(),
-        });
-    }
-
-    /// Subscribe to `astrid.v1.elicit`, wait for the request, return its
-    /// `request_id` AND the principal stamped on the request message. The host
-    /// mints the id internally, so a driver learns it the same way every real
-    /// answerer does — off the published request.
-    async fn await_request(mut req_rx: astrid_events::EventReceiver) -> (Uuid, Option<String>) {
-        loop {
-            let ev = tokio::time::timeout(Duration::from_secs(5), req_rx.recv())
-                .await
-                .expect("elicit request observed")
-                .expect("bus open");
-            if let AstridEvent::Ipc { message, .. } = &*ev
-                && let IpcPayload::ElicitRequest { request_id, .. } = &message.payload
-            {
-                return (*request_id, message.principal.clone());
-            }
-        }
-    }
-
-    /// End-to-end: a cross-principal reply must be ignored; the matching reply
-    /// (published AFTER the bad one) unblocks the waiter with the right value.
-    /// Fails if the principal check is removed — the waiter would unblock on
-    /// the wrong-principal reply and return "intruder".
-    ///
-    /// Test 4 folded in: also asserts the published `ElicitRequest` carries the
-    /// originating principal, pinning that the host STAMPS it on the outbound
-    /// request (not just checks it on replies).
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn cross_principal_reply_does_not_unblock_matching_does() {
-        let rt = tokio::runtime::Handle::current();
-        let mut state = minimal_host_state(rt);
-        state.principal = astrid_core::PrincipalId::new("agent-alice").unwrap();
-        state.lifecycle_phase = Some(LifecyclePhase::Install);
-
-        let bus = state.event_bus.clone();
-        let req_rx = bus.subscribe_topic("astrid.v1.elicit");
-
-        // Drive the blocking host fn on a dedicated thread (it uses
-        // block_in_place). We can't move the non-Send HostState across an
-        // await, so run it inside spawn_blocking and bridge the result back.
-        let elicit_handle =
-            tokio::task::spawn_blocking(move || (state.elicit(text_request("api_url")), state));
-
-        // Learn the request_id, then publish a WRONG-principal reply first.
-        let (request_id, req_principal) = await_request(req_rx).await;
-        assert_eq!(
-            req_principal.as_deref(),
-            Some("agent-alice"),
-            "host must stamp the originating principal on the outbound ElicitRequest"
-        );
-        publish_response(&bus, request_id, "agent-bob", "intruder");
-
-        // Give the loop a beat to (correctly) reject the intruder reply, then
-        // send the legitimate one.
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        publish_response(&bus, request_id, "agent-alice", "legit");
-
-        let (result, _state) = elicit_handle.await.expect("elicit thread joined");
-        match result {
-            Ok(ElicitResponse::Value(v)) => {
-                assert_eq!(v, "legit", "must return the matching principal's value");
-                assert_ne!(v, "intruder", "cross-principal reply must not win");
-            },
-            other => panic!("expected matching value, got {other:?}"),
-        }
-    }
-
-    /// Test 2 — CANCEL TOKEN UNBLOCKS THE WAIT: firing the host state's
-    /// `cancel_token` mid-wait (no reply published) must make `elicit()` return
-    /// promptly, well under `MAX_ELICIT_TIMEOUT_MS`, rather than waiting the full
-    /// 120s. Guards that the deadline-loop restructure kept unload-wins: the
-    /// outer `bounded_block_on_cancellable` races the token against the wait.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn cancel_token_unblocks_elicit_wait() {
-        let rt = tokio::runtime::Handle::current();
-        let mut state = minimal_host_state(rt);
-        state.principal = astrid_core::PrincipalId::new("agent-alice").unwrap();
-        state.lifecycle_phase = Some(LifecyclePhase::Install);
-
-        let bus = state.event_bus.clone();
-        let cancel_token = state.cancel_token.clone();
-        let req_rx = bus.subscribe_topic("astrid.v1.elicit");
-
-        let start = std::time::Instant::now();
-        let elicit_handle =
-            tokio::task::spawn_blocking(move || (state.elicit(text_request("api_url")), state));
-
-        // Wait until the host has published its request (so it is actually
-        // blocked in the wait), then cancel — NO reply is ever published.
-        let (_request_id, _principal) = await_request(req_rx).await;
-        cancel_token.cancel();
-
-        let (result, _state) = elicit_handle.await.expect("elicit thread joined");
-        let elapsed = start.elapsed();
-
-        // Cancellation surfaces as Timeout (the helper returns None for both
-        // deadline-expiry and cancel — same as before the refactor).
-        assert!(
-            matches!(result, Err(ErrorCode::Timeout)),
-            "cancelled wait must return Timeout, got {result:?}"
-        );
-        // The load-bearing property: it returned PROMPTLY, not after 120s.
-        assert!(
-            elapsed < Duration::from_secs(2),
-            "cancel must unblock promptly, took {elapsed:?}"
-        );
-    }
-
-    /// Test 3 — FULL-HOST-FN CANCELLATION ANSWER: a MATCHING-principal reply
-    /// carrying both `value: None` and `values: None` (the user-cancel sentinel)
-    /// must drive the real `elicit()` to `Err(ErrorCode::Cancelled)` — not
-    /// `Timeout`, not `Ok`. Exercises the cancel-sentinel mapping through the
-    /// whole host fn, past the principal check.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn matching_cancel_sentinel_returns_cancelled() {
-        let rt = tokio::runtime::Handle::current();
-        let mut state = minimal_host_state(rt);
-        state.principal = astrid_core::PrincipalId::new("agent-alice").unwrap();
-        state.lifecycle_phase = Some(LifecyclePhase::Install);
-
-        let bus = state.event_bus.clone();
-        let req_rx = bus.subscribe_topic("astrid.v1.elicit");
-
-        let elicit_handle =
-            tokio::task::spawn_blocking(move || (state.elicit(text_request("api_url")), state));
-
-        let (request_id, _principal) = await_request(req_rx).await;
-        publish_cancel(&bus, request_id, "agent-alice");
-
-        let (result, _state) = elicit_handle.await.expect("elicit thread joined");
-        assert!(
-            matches!(result, Err(ErrorCode::Cancelled)),
-            "both-None reply from the matching principal must map to Cancelled, got {result:?}"
-        );
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Chain tests: drive `has_secret` synchronously on a HostState with manually-
-// installed invocation fields. Verifies `effective_secret_store()` wiring: a
-// key set via the invocation store must not be visible via the load-time
-// store and vice versa. Mirrors the pattern established in `host/fs.rs` for
-// per-invocation VFS re-scoping (#549).
-// ---------------------------------------------------------------------------
 #[cfg(test)]
-mod secret_chain_tests {
-    use std::sync::Arc;
-
-    use crate::engine::wasm::bindings::astrid::elicit::host::Host as ElicitHost;
-    use crate::engine::wasm::host_state::HostState;
-    use crate::engine::wasm::test_fixtures::{mem_secret_store, minimal_host_state};
-    use astrid_storage::secret::SecretStore;
-
-    /// Build a HostState whose load-time `secret_store` points at a fresh,
-    /// namespace-isolated KV-backed store. Returns the state and an `Arc`
-    /// handle to that owner store so tests can seed secrets through it.
-    fn make_host_state_with_secret(
-        rt: tokio::runtime::Handle,
-        owner_namespace: &str,
-    ) -> (HostState, Arc<dyn SecretStore>) {
-        let owner_secret = mem_secret_store(owner_namespace, rt.clone());
-        let mut state = minimal_host_state(rt);
-        state.secret_store = Arc::clone(&owner_secret);
-        (state, owner_secret)
-    }
-
-    /// Fresh invocation-scoped secret store (principal-namespaced in real
-    /// `invoke_interceptor`; arbitrary in tests).
-    fn make_invocation_store(rt: tokio::runtime::Handle, namespace: &str) -> Arc<dyn SecretStore> {
-        mem_secret_store(namespace, rt)
-    }
-
-    /// Drive a closure in a blocking context so KvSecretStore's internal
-    /// `Handle::block_on` works — same sync/async bridge pattern as
-    /// production host functions.
-    async fn blocking<T, F>(f: F) -> T
-    where
-        T: Send + 'static,
-        F: FnOnce() -> T + Send + 'static,
-    {
-        tokio::task::spawn_blocking(f)
-            .await
-            .expect("spawn_blocking join")
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn has_secret_reads_invocation_store_when_installed() {
-        let rt = tokio::runtime::Handle::current();
-        let (mut state, owner_secret) =
-            make_host_state_with_secret(rt.clone(), "capsule:test-owner");
-        let alice_secret = make_invocation_store(rt, "capsule:test-alice");
-
-        // Owner has `shared_key`; Alice does not.
-        {
-            let s = Arc::clone(&owner_secret);
-            blocking(move || s.set("shared_key", "owner-val").unwrap()).await;
-        }
-        state.invocation_secret_store = Some(Arc::clone(&alice_secret));
-
-        // Via the accessor, `has_secret` consults Alice's store — the owner's
-        // entry is not visible.
-        let (state, got) = blocking(move || {
-            let mut s = state;
-            let got = s.has_secret("shared_key".to_string()).unwrap();
-            (s, got)
-        })
-        .await;
-        assert!(!got, "invocation store is empty; owner's key must not leak");
-
-        // Alice sets her own; owner's view is unchanged.
-        {
-            let s = Arc::clone(&alice_secret);
-            blocking(move || s.set("shared_key", "alice-val").unwrap()).await;
-        }
-        let (mut state, got) = blocking(move || {
-            let mut s = state;
-            let got = s.has_secret("shared_key".to_string()).unwrap();
-            (s, got)
-        })
-        .await;
-        assert!(got);
-
-        // Drop invocation context: falls back to owner's store.
-        state.invocation_secret_store = None;
-        let (_state, got) = blocking(move || {
-            let mut s = state;
-            let got = s.has_secret("shared_key".to_string()).unwrap();
-            (s, got)
-        })
-        .await;
-        assert!(got, "owner's key still present after clear");
-
-        // Sanity: owner never saw Alice's value.
-        let (owner_val, alice_val) = blocking(move || {
-            (
-                owner_secret.get("shared_key").unwrap(),
-                alice_secret.get("shared_key").unwrap(),
-            )
-        })
-        .await;
-        assert_eq!(owner_val.as_deref(), Some("owner-val"));
-        assert_eq!(alice_val.as_deref(), Some("alice-val"));
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn has_secret_falls_back_to_load_time_store() {
-        // Regression guard: single-tenant path (no invocation store installed)
-        // must see load-time secrets.
-        let rt = tokio::runtime::Handle::current();
-        let (state, owner_secret) = make_host_state_with_secret(rt, "capsule:test-owner");
-        {
-            let s = Arc::clone(&owner_secret);
-            blocking(move || s.set("api_key", "sk-load").unwrap()).await;
-        }
-        assert!(state.invocation_secret_store.is_none());
-        let (_state, got1, got2) = blocking(move || {
-            let mut state = state;
-            let got1 = state.has_secret("api_key".to_string()).unwrap();
-            let got2 = state.has_secret("other_key".to_string()).unwrap();
-            (state, got1, got2)
-        })
-        .await;
-        assert!(got1);
-        assert!(!got2);
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn has_secret_isolates_across_sequential_invocations() {
-        // Same HostState, invocation store swapped between calls — each call
-        // sees only the currently-installed principal's secrets.
-        let rt = tokio::runtime::Handle::current();
-        let (mut state, _owner_secret) =
-            make_host_state_with_secret(rt.clone(), "capsule:test-owner");
-
-        let alice_secret = make_invocation_store(rt.clone(), "capsule:test-alice");
-        let bob_secret = make_invocation_store(rt, "capsule:test-bob");
-        {
-            let a = Arc::clone(&alice_secret);
-            let b = Arc::clone(&bob_secret);
-            blocking(move || {
-                a.set("pk", "alice-pk").unwrap();
-                b.set("pk", "bob-pk").unwrap();
-            })
-            .await;
-        }
-
-        state.invocation_secret_store = Some(Arc::clone(&alice_secret));
-        let (mut state, alice_view) = blocking(move || {
-            let mut s = state;
-            let v = s.has_secret("pk".to_string()).unwrap();
-            (s, v)
-        })
-        .await;
-        assert!(alice_view);
-        state.invocation_secret_store = None;
-
-        state.invocation_secret_store = Some(Arc::clone(&bob_secret));
-        let (_state, bob_view) = blocking(move || {
-            let mut s = state;
-            let v = s.has_secret("pk".to_string()).unwrap();
-            (s, v)
-        })
-        .await;
-        assert!(bob_view);
-
-        // Both isolated: each only sees its own key.
-        let (a_val, b_val) = blocking(move || {
-            (
-                alice_secret.get("pk").unwrap(),
-                bob_secret.get("pk").unwrap(),
-            )
-        })
-        .await;
-        assert_eq!(a_val.as_deref(), Some("alice-pk"));
-        assert_eq!(b_val.as_deref(), Some("bob-pk"));
-    }
-}
+#[path = "elicit/secret_chain_tests.rs"]
+mod secret_chain_tests;
