+++ scripts/e2e/runtime-adversarial-smoke.sh
@@ -216,6 +216,55 @@ PY
   printf '%s\n' "$request_id"
 }
 
+run_agent_request_http_guardrails() {
+  local user_bearer=$1
+  local user_principal=$2
+  local status
+  local python_bin="${PYTHON:-python3}"
+
+  note "checking elicit-response HTTP guardrails"
+  local elicit_request_id
+  elicit_request_id="$("$python_bin" -c 'import uuid; print(uuid.uuid4())')"
+  status="$(http_status GET /api/agent/requests "" "" "$ARTIFACTS/agent-requests-unauth.json")"
+  assert_status "unauthenticated agent request stream denied" "$status" 401
+  curl -sN --max-time 3 \
+    -H "Authorization: Bearer $user_bearer" \
+    "$GATEWAY/api/agent/requests" \
+    > "$ARTIFACTS/agent-requests.sse" 2>&1 || true
+  grep -q '^event: ready' "$ARTIFACTS/agent-requests.sse" \
+    || fail "agent request stream did not emit ready event"
+  grep -q "\"principal\":\"$user_principal\"" "$ARTIFACTS/agent-requests.sse" \
+    || fail "agent request stream ready event did not carry caller principal"
+  status="$(http_status POST /api/agent/elicit-response "" \
+    "{\"request_id\":\"$elicit_request_id\",\"value\":\"unauth\"}" \
+    "$ARTIFACTS/elicit-response-unauth.json")"
+  assert_status "unauthenticated elicit response denied" "$status" 401
+  status="$(http_status POST /api/agent/elicit-response "$user_bearer" \
+    "{\"request_id\":\"$elicit_request_id\",\"value\":\"scalar\",\"values\":[\"array\"]}" \
+    "$ARTIFACTS/elicit-response-ambiguous.json")"
+  assert_status "ambiguous elicit response rejected" "$status" 400
+  status="$(http_status POST /api/agent/elicit-response "$user_bearer" \
+    "{\"request_id\":\"$elicit_request_id\",\"value\":\"answered\",\"principal\":\"default\"}" \
+    "$ARTIFACTS/elicit-response-spoofed-principal.json")"
+  assert_status "elicit response body principal ignored" "$status" 202
+  status="$(http_status POST /api/agent/elicit-response "$user_bearer" \
+    "{\"request_id\":\"$elicit_request_id\"}" \
+    "$ARTIFACTS/elicit-response-cancel.json")"
+  assert_status "elicit response cancel sentinel accepted" "$status" 202
+  status="$(http_status POST /api/agent/approval-response "" \
+    '{"request_id":"approval-unauth","decision":"approve"}' \
+    "$ARTIFACTS/approval-response-unauth.json")"
+  assert_status "unauthenticated approval response denied" "$status" 401
+  status="$(http_status POST /api/agent/approval-response "$user_bearer" \
+    '{"request_id":"approval-bad","decision":"maybe"}' \
+    "$ARTIFACTS/approval-response-invalid.json")"
+  assert_status "invalid approval response rejected" "$status" 400
+  status="$(http_status POST /api/agent/approval-response "$user_bearer" \
+    '{"request_id":"approval-spoof","decision":"deny","principal":"default"}' \
+    "$ARTIFACTS/approval-response-spoofed-principal.json")"
+  assert_status "approval response body principal ignored" "$status" 202
+}
+
 assert_artifact_lacks_text() {
   local path=$1
   local text=$2
